% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataTypesProteo.R
\name{compareDataTypesProteo}
\alias{compareDataTypesProteo}
\title{Compare Proteo}
\usage{
compareDataTypesProteo(
  tmt_lsdt,
  silac_lsdt,
  geneCol_v = "Gene",
  col_v = "diffExp",
  levels_lsv = list(No = "NO", Up = c("upHigh", "upMed", "upLow"), Down = c("downHigh",
    "downMed", "downLow"))
)
}
\arguments{
\item{tmt_lsdt}{list of tmt results. Must be run through classifyProteo first.}

\item{silac_lsdt}{list of silac results. Must be run through classifyProteo first.}

\item{geneCol_v}{column name in both data.tables that refers to the genes}

\item{col_v}{column name in both data.tables that refers to differential expression "levels", from classifyProteo}

\item{levels_lsv}{levels created from classifyProteo. This shouldn't change, but could theoretically.}
}
\value{
list
}
\description{
Compare the significance groups of proteomics results
between tmt and silac
}
\details{
Want to get an idea of what to expect in each experiment. How many genes significant/shared/etc.
}
