% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePhenotypesProteo.R
\name{comparePhenotypesProteo}
\alias{comparePhenotypesProteo}
\title{Compare Proteo}
\usage{
comparePhenotypesProteo(
  data1_dt,
  data2_dt = NULL,
  names_v = NULL,
  geneCol_v = "Gene",
  col_v = "diffExp",
  levels_lsv = list(No = "NO", Up = c("upHigh", "upMed", "upLow"), Down = c("downHigh",
    "downMed", "downLow"))
)
}
\arguments{
\item{data1_dt}{either a list of two data.tables, or a single data.table}

\item{data2_dt}{if data1_dt is a list, this will be ignored. If it's a data.table, this must be a data.table}

\item{names_v}{optional vector containing the experiment names for data1 and data2. Only used if data1_dt is NOT a list.}

\item{geneCol_v}{column name in both data.tables that refers to the genes}

\item{col_v}{column name in both data.tables that refers to differential expression "levels", from classifyProteo}

\item{levels_lsv}{levels created from classifyProteo. This shouldn't change, but could theoretically.}
}
\value{
list
}
\description{
Compare the significance groups of proteomics results
between two different experiments from the same data type
}
\details{
Want to get an idea of what to expect in each experiment. How many genes significant/shared/etc.
}
