% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareProteo.R
\name{compareProteo}
\alias{compareProteo}
\title{Compare Proteo}
\usage{
compareProteo(
  tmt_lsdt,
  silac_lsdt,
  geneCol_v = "Gene",
  col_v = "diffExp",
  levels_lsv = list(No = "NO", Up = c("upHigh", "upMed", "upLow"), Down = c("downHigh",
    "downMed", "downLow")),
  markers_lsdt
)
}
\arguments{
\item{tmt_lsdt}{list of tmt results. Must be run through classifyProteo first.}

\item{silac_lsdt}{list of silac results. Must be run through classifyProteo first.}

\item{levels_lsv}{levels created from classifyProteo. This shouldn't change, but could theoretically.}

\item{markers_lsdt}{list of data.tables containing markers to use. Must have at least a 'Gene' column}
}
\value{
list
}
\description{
Compare the significance groups of proteomics results
between tmt and silac
}
\details{
Want to get an idea of what to expect in each experiment. How many genes significant/shared/etc.
}
