% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneGroups.R
\name{getGeneGroups}
\alias{getGeneGroups}
\title{Get Gene Groups}
\usage{
getGeneGroups(
  data_dt,
  levels_lsv,
  removeNA_v = T,
  col_v = "diffExp",
  geneCol_v = "Gene"
)
}
\arguments{
\item{data_dt}{tmt or silace results table}

\item{levels_lsv}{the significance levels}

\item{removeNA_v}{logical. Should 'na' genes be removed?}
}
\value{
list of lists
}
\description{
Get all the genes in different signficance groups
}
