% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerBarPlot.R
\name{markerBarPlot}
\alias{markerBarPlot}
\title{Marker Bar Plot}
\usage{
markerBarPlot(
  data_lsdt,
  geneCol_v = "Gene",
  y_v = "logFC",
  fillCol_v = "Class",
  colors_v = NULL,
  markers_dt,
  markerName_v,
  expName_v = NULL,
  title_v = NULL
)
}
\arguments{
\item{geneCol_v}{column in data_dt containing gene names}

\item{y_v}{column in data_dt that has the y variable}

\item{fillCol_v}{variable to color bars with}

\item{markers_dt}{data.table of markers to check for in data_dt. must have geneCol_v}

\item{markerName_v}{name of markers to include in plot}

\item{expName_v}{name of phenotype to include in plot. Only used if data_lsdt is a single data.table, then it is required.}

\item{title_v}{optional character vector to prepend to the title.}

\item{data_dt}{list of data.tables of proteomics DE results (can also just be one table)}
}
\value{
ggplot
}
\description{
make a barplot displaying expression values for specific markers
}
