% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseProteoList.R
\name{parseProteoList}
\alias{parseProteoList}
\title{Parse proteo list}
\usage{
parseProteoList(list_lslsv)
}
\arguments{
\item{list_lslsv}{gene groups of tmt or silac results}
}
\value{
summary stats
}
\description{
get summary stats
}
\details{
list_lslsv must have 3 elements. Each is a list:
\enumerate{
\item "No" - 'NO'
\item "Up" - 'upHigh', 'upMed', 'upLow'
\item "Down" - 'downHigh', 'downMed', 'downLow'
}
}
