% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompareDataTypesProteo.R
\name{plotCompareDataTypesProteo}
\alias{plotCompareDataTypesProteo}
\title{Plot Compare Data Types Proteo}
\usage{
plotCompareDataTypesProteo(
  tmt_dt,
  silac_dt,
  name_v,
  tmtLFCcol_v = "logFC",
  tmtPvalCol_v = "FDR",
  silacLFCcol_v = "Z-score",
  silacPvalCol_v = "BH_adjusted",
  silacVolcCol_v = "Log2(B/A)",
  lfc_v = 0.5,
  pval_v = c(0.01, 0.05)
)
}
\arguments{
\item{tmt_dt}{data.table of tmt results}

\item{silac_dt}{data.table of silac results}

\item{name_v}{Name of the experiment (e.g. M1E)}

\item{tmtLFCcol_v}{column name in tmt_dt that corresponds to log fold change}

\item{tmtPvalCol_v}{column name in tmt_dt that corresponds to adjusted p-value}

\item{silacLFCcol_v}{column name in silac_dt that corresponds to log fold change}

\item{silacPvalCol_v}{column name in silac_dt that corresponds to adjusted p-value}

\item{lfc_v}{log-fold change cut-off value}

\item{pval_v}{p-value cut-offs for grouping.}

\item{silacVolCol_v}{not sure if lfcol or this col should be plotted for silac}
}
\value{
volcano plot, venn diagrams, scatterplot
}
\description{
Make various plots to compare the genes returned by each data type
}
