% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparePhenotypesProteo.R
\name{plotComparePhenotypesProteo}
\alias{plotComparePhenotypesProteo}
\title{Plot Compare Phenotypes Proteo}
\usage{
plotComparePhenotypesProteo(
  data1_dt,
  data2_dt = NULL,
  names_v = NULL,
  lfcCol1_v = "logFC",
  pValCol1_v = "FDR",
  lfcCol2_v = "logFC",
  pValCol2_v = "FDR",
  lfc_v = 0.5,
  pval_v = c(0.01, 0.05)
)
}
\arguments{
\item{data1_dt}{either a list of two data.tables, or a single data.table}

\item{data2_dt}{if data1_dt is a list, this will be ignored. If it's a data.table, this must be a data.table}

\item{names_v}{optional vector containing the experiment names for data1 and data2. Only used if data1_dt is NOT a list.}

\item{lfcCol1_v}{column name in first data.table that corresponds to log fold change}

\item{pValCol1_v}{column name in first data.table that corresponds to adjusted p-value}

\item{lfcCol2_v}{column name in second data.table that corresponds to log fold change}

\item{pValCol2_v}{column name in second data.table that corresponds to adjusted p-value}

\item{lfc_v}{log-fold change cut-off value}

\item{pval_v}{p-value cut-offs for grouping.}
}
\value{
volcano plot, venn diagrams, scatterplot
}
\description{
Make various plots to compare the genes returned by two phenotypes
}
