% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Run GSEA}
\usage{
plotGSEA(
  data_dt,
  list_v,
  listName_v,
  title_v = NULL,
  name_v,
  lfcCol_v = "logFC",
  le_v = F,
  shortenNames_v = T,
  lfc_v = 0.5,
  pval_v = 0.05,
  labelSize_v = 1
)
}
\arguments{
\item{data_dt}{DEG table to run GSEA on}

\item{list_v}{list of genes (gene set) used for GSEA}

\item{listName_v}{name of the gene set}

\item{title_v}{optional plot title}

\item{name_v}{name of experiment (e.g. M1E)}

\item{le_v}{logical indicating if leading edge should be run. DOESN'T WORK RIGHT NOW.}

\item{shortenNames_v}{logical indicating if gene set names should be shortened.}

\item{lfc_v}{log fold change - passed through for leading edge gene filter}

\item{pval_v}{adjusted p-value - used to filter significant pathways and also passed through for leading edge filter}

\item{labelSize_v}{how big to make leading edge volcano label text. Default is 1.}

\item{pvalCol_v}{column in data_dt that has p-values}

\item{plot_v}{logical indicating if plots should be made/saved or just the results calculated.}
}
\value{
not sure yet
}
\description{
Run GSEA and output results. Optionally output plots
}
