% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoVolcano.R
\name{proteoVolcano}
\alias{proteoVolcano}
\title{ProteoVolcano}
\usage{
proteoVolcano(
  data_dt,
  geneCol_v = "Gene",
  lfcCol_v,
  pvalCol_v,
  lfc_v = 0.5,
  ident1_v,
  labelAll_v = F,
  labelTop_v = 20,
  labelDir_v = "both",
  labelSize_v = 1,
  title_v
)
}
\arguments{
\item{data_dt}{data.table of protein expression}

\item{geneCol_v}{column in data_dt identifying the gene}

\item{lfcCol_v}{column in data_dt indicating the fold change results (logFC for TMT; Z-score for Silac)}

\item{pvalCol_v}{column in data_dt indicating the significance value (FDR for TMT; BH_adjusted for Silac)}

\item{lfc_v}{value for cut-off for fold change. Will be included in plot as vertical dotted line}

\item{ident1_v}{name of the reference treatment. Genes on right of plot are higher in this treatment than the other}

\item{labelAll_v}{logical indication}

\item{labelTop_v}{optional, either NULL or a number indicating top N genes to label.}

\item{labelDir_v}{can be "up", "down", or "both", indicating with side(s) of the volcano to label.}

\item{labelSize_v}{pch size}

\item{title_v}{what to name the plot}
}
\value{
volcano ggplot
}
\description{
Volcano plot for proteomics data (TMT and Silac)
low sig: 0.1 > p >= 0.05
medium sig: 0.05 > p >= 0.01
high sig: 0.01 > p
}
\details{
Fill this in later
}
