% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeProteo.R
\name{summarizeProteo}
\alias{summarizeProteo}
\title{Summarize Proteo}
\usage{
summarizeProteo(
  data_lsdt,
  geneCol_v = "Gene",
  col_v = "diffExp",
  levels_lsv = list(No = "NO", Up = c("upHigh", "upMed", "upLow"), Down = c("downHigh",
    "downMed", "downLow"))
)
}
\arguments{
\item{data_lsdt}{list of tmt or silac datasets to summarize}

\item{geneCol_v}{column name in the data.tables that indicates the gene}

\item{col_v}{column to use to summarize}

\item{levels_lsv}{the up and down levels that are in col_v}
}
\value{
summary table
}
\description{
summarize proteomics results
}
\details{
Summarize number of up/down genes by confidence group.
Both input lists need to have been run through classifyProteo() prior.
}
