% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerBarPlot.R
\name{markerBarPlot}
\alias{markerBarPlot}
\title{Marker Bar Plot}
\usage{
markerBarPlot(
  data_dt,
  geneCol_v = "Gene",
  y_v = "logFC",
  fillCol_v = "Class",
  colors_v = NULL,
  markers_dt,
  markerName_v,
  expName_v
)
}
\arguments{
\item{data_dt}{data.table of proteomics DE results}

\item{geneCol_v}{column in data_dt containing gene names}

\item{y_v}{column in data_dt that has the y variable}

\item{fillCol_v}{variable to color bars with}

\item{markers_dt}{data.table of markers to check for in data_dt. must have geneCol_v}

\item{markerName_v}{name of markers to include in plot}

\item{expName_v}{name of phenotype to include in plot}
}
\value{
ggplot
}
\description{
make a barplot displaying expression values for specific markers
}
